# docs/swagger.yaml
basePath: /
definitions:
  GinProject_internal_model.ApiUsers:
    properties:
      account:
        type: string
      created_at:
        type: string
      fullname:
        type: string
      updated_at:
        type: string
    type: object
  GinProject_internal_model.BasicSuccessResponse:
    properties:
      error:
        type: string
    type: object
  GinProject_internal_model.CreateUserData:
    properties:
      account:
        maxLength: 50
        minLength: 4
        type: string
      fullname:
        maxLength: 50
        minLength: 4
        type: string
      password:
        maxLength: 50
        minLength: 8
        type: string
    type: object
  GinProject_internal_model.LoginSuccessResponse:
    properties:
      error:
        type: string
      token:
        type: string
    type: object
  GinProject_internal_model.LoginUserData:
    properties:
      account:
        maxLength: 50
        minLength: 4
        type: string
      password:
        maxLength: 50
        minLength: 8
        type: string
    type: object
  GinProject_internal_model.SuccessUserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/GinProject_internal_model.ApiUsers'
        type: array
      error:
        type: string
    type: object
  GinProject_internal_model.SuccessUserResponse:
    properties:
      data:
        $ref: '#/definitions/GinProject_internal_model.ApiUsers'
      error:
        type: string
    type: object
  GinProject_internal_model.UpdateUserData:
    properties:
      fullname:
        maxLength: 50
        minLength: 4
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Johnny.Chen
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GinProject
  version: "1.0"
paths:
  /login:
    post:
      parameters:
      - description: user data
        in: body
        name: userdata
        required: true
        schema:
          $ref: '#/definitions/GinProject_internal_model.LoginUserData'
      produces:
      - application/json
      responses:
        "200":
          description: return jwt token
          schema:
            $ref: '#/definitions/GinProject_internal_model.LoginSuccessResponse'
      summary: user login
      tags:
      - user
  /user:
    get:
      parameters:
      - description: page for pagination
        in: query
        name: page
        type: integer
      - description: page size for pagination
        in: query
        name: pagesize
        type: integer
      - description: find user with fullname
        in: query
        name: fullname
        type: string
      - description: sort result by column
        in: query
        name: sortby
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: return user list
          schema:
            $ref: '#/definitions/GinProject_internal_model.SuccessUserListResponse'
      summary: get user list
      tags:
      - user
    post:
      parameters:
      - description: user data
        in: body
        name: userdata
        required: true
        schema:
          $ref: '#/definitions/GinProject_internal_model.CreateUserData'
      produces:
      - application/json
      responses:
        "200":
          description: return error message
          schema:
            $ref: '#/definitions/GinProject_internal_model.BasicSuccessResponse'
      summary: create user
      tags:
      - user
  /user/{account}:
    delete:
      parameters:
      - description: user account
        in: path
        name: account
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'return error message '
          schema:
            $ref: '#/definitions/GinProject_internal_model.BasicSuccessResponse'
      summary: delete user
      tags:
      - user
    get:
      parameters:
      - description: page for pagination
        in: query
        name: page
        type: integer
      - description: page size for pagination
        in: query
        name: pagesize
        type: integer
      - description: find user with fullname
        in: query
        name: fullname
        type: string
      - description: sort result by column
        in: query
        name: sortby
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: return user
          schema:
            $ref: '#/definitions/GinProject_internal_model.SuccessUserResponse'
      summary: get user
      tags:
      - user
    patch:
      parameters:
      - description: user account
        in: path
        name: account
        required: true
        type: string
      - description: user data
        in: body
        name: userdata
        required: true
        schema:
          $ref: '#/definitions/GinProject_internal_model.UpdateUserData'
      produces:
      - application/json
      responses:
        "200":
          description: 'return error message '
          schema:
            $ref: '#/definitions/GinProject_internal_model.BasicSuccessResponse'
      summary: update user
      tags:
      - user
schemes:
- http
swagger: "2.0"